# api/swagger/openapi.json
{
  "components": {
    "schemas": {
...
      "ClusterInfoPeerResponse": {
        "properties": {
          "active": {
            "type": "integer"
          },
          "current": {
            "type": "boolean"
          },
          "lag": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "offline": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ClusterInfoResponse": {
        "properties": {
          "cluster_name": {
            "type": "string"
          },
          "cluster_url": {
            "type": "string"
          },
          "is_clustered": {
            "type": "boolean"
          },
          "jetstream": {
            "$ref": "#/components/schemas/ClusterJetStreamInfo"
          },
          "server_name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ClusterJetStreamInfo": {
        "properties": {
          "api_level": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "tier": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ClusterPlacementResponse": {
        "properties": {
          "cluster": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ClusterStreamClusterResponse": {
        "properties": {
          "leader": {
            "type": "string"
          },
          "leader_since": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "raft_group": {
            "type": "string"
          },
          "replicas": {
            "items": {
              "$ref": "#/components/schemas/ClusterInfoPeerResponse"
            },
            "type": "array"
          },
          "system_account": {
            "type": "boolean"
          },
          "traffic_account": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ClusterStreamReplicaResponse": {
        "properties": {
          "cluster": {
            "$ref": "#/components/schemas/ClusterStreamClusterResponse"
          },
          "is_clustered": {
            "type": "boolean"
          },
          "mirror": {
            "$ref": "#/components/schemas/ClusterStreamSource"
          },
          "placement": {
            "$ref": "#/components/schemas/ClusterPlacementResponse"
          },
          "replicas": {
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ClusterStreamSource"
            },
            "type": "array"
          },
          "stream": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ClusterStreamSource": {
        "properties": {
          "domain": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "NATS Horizon API",
    "version": "1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/cluster/info": {
      "get": {
        "description": "Returns JetStream cluster topology and server information",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterInfoResponse"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get cluster information",
        "tags": [
          "cluster"
        ]
      }
    },
...
    "/cluster/streams/{name}/replicas": {
      "get": {
        "description": "Returns replication, mirror, source, and cluster placement info for a stream",
        "parameters": [
          {
            "description": "Stream name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterStreamReplicaResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get stream replicas",
        "tags": [
          "cluster"
        ]
      }
    },
    "/config": {
      "get": {
        "description": "Returns the current server configuration",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get current configuration",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Updates server configuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "description": "Configuration update request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update configuration",
        "tags": [
          "config"
        ]
      }
    },
...
    "/config/setup/complete": {
      "post": {
        "description": "Marks the initial setup as completed",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Complete setup",
        "tags": [
          "config"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}